Sound Module Update
FIRST: I updated the sound module. It is no longer a class, and has two
functions. Import 'sounds', and then simply  type
sounds.play('NameOfSound') and it will play that sound effect. I also
have a function 'loop' with the same input for background
music/ambience. This should be much cleaner code, and all of the
directory and file name information is kept in the sounds.py file.

I also added test sounds to the background ambience, and the rifle.

NOTE: USE WAV FILES, MP3 DOES NOT WORK WITH THE 'SOUND' MIXER, ONLY WITH
'MUSIC'. I need to go back and redownload the wav versions of the other
files. Feel free to add sound files to appropriate folders if you find
files that make sense.